@@ -4,7 +4,7 @@
     "name": "VampireChicken12"
   },
   "displayName": "YouTube Enhancer",
-  "version": "1.26.0",
+  "version": "1.27.0",
   "description": "YouTube Enhancer is a simple extension that adds some useful features to YouTube.",
   "license": "MIT",
   "repository": {
@@ -28,8 +28,9 @@
     "@monaco-editor/react": "^4.6.0",
     "@tanstack/react-query": "^5.18.0",
     "dotenv": "^16.3.1",
+    "globals": "^15.11.0",
     "i18next": "^23.7.3",
-    "monaco-editor": "^0.50.0",
+    "monaco-editor": "^0.52.0",
     "react": "^18.2.0",
     "react-colorful": "^5.6.1",
     "react-dom": "^18.2.0",
@@ -41,35 +42,38 @@
     "webextension-polyfill": "^0.12.0"
   },
   "devDependencies": {
+    "@eslint/compat": "^1.2.0",
+    "@eslint/js": "^9.12.0",
     "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/exec": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
     "@thedutchcoder/postcss-rem-to-px": "^0.0.2",
-    "@total-typescript/ts-reset": "^0.5.1",
+    "@total-typescript/ts-reset": "^0.6.0",
     "@types/archiver": "^6.0.1",
-    "@types/chrome": "^0.0.268",
-    "@types/node": "^20.9.0",
+    "@types/chrome": "^0.0.280",
+    "@types/eslint__js": "^8.42.3",
+    "@types/node": "^22.1.0",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
-    "@types/webextension-polyfill": "^0.10.6",
+    "@types/webextension-polyfill": "^0.12.1",
     "@types/youtube-player": "^5.5.10",
-    "@typescript-eslint/eslint-plugin": "^7.0.0",
-    "@typescript-eslint/parser": "^7.7.0",
+    "@typescript-eslint/eslint-plugin": "^8.8.1",
+    "@typescript-eslint/parser": "^8.8.1",
     "@vitejs/plugin-react-swc": "^3.4.1",
     "archiver": "^7.0.1",
     "autoprefixer": "^10.4.16",
     "clsx": "^2.0.0",
     "concurrently": "^8.2.2",
-    "eslint": "^8.53.0",
+    "eslint": "^9.12.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-no-secrets": "^1.0.2",
-    "eslint-plugin-perfectionist": "^2.3.0",
-    "eslint-plugin-prettier": "^5.0.1",
-    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-perfectionist": "^3.8.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-promise": "^7.1.0",
     "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-tailwindcss": "^3.13.0",
     "fs-extra": "^11.1.1",
     "get-installed-browsers": "^0.1.7",
@@ -81,8 +85,9 @@
     "tailwindcss": "^3.3.6",
     "ts-json-as-const": "^1.0.7",
     "ts-node": "^10.9.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.0.12",
+    "typescript": "5.6",
+    "typescript-eslint": "^8.8.1",
+    "vite": "^5.4.8",
     "zod": "^3.22.4",
     "zod-error": "^1.5.0"
   }
