@@ -1,7 +1,7 @@
 import type { i18nInstanceType } from "./i18n";
 
 declare module "*.svg" {
-	import React = require("react");
+	import React from "react";
 	export const ReactComponent: React.SFC<React.SVGProps<SVGSVGElement>>;
 	const src: string;
 	export default src;
@@ -58,7 +58,9 @@ declare module "node_modules/@types/youtube-player/dist/types" {
 		getProgressState(): ProgressState;
 		getVideoBytesLoaded(): Promise<number>;
 		getVideoData(): Promise<VideoData>;
+		loadModule(moduleName: string): void;
 		setPlaybackQualityRange(suggestedQuality: string): Promise<void>;
+		unloadModule(moduleName: string): void;
 	}
 }
 declare global {
@@ -68,6 +70,7 @@ declare global {
 	}
 	interface Window {
 		audioCtx: AudioContext;
+		cachedPlaylistDuration: { playlistId: string; totalTimeSeconds: number } | null;
 		gainNode: GainNode;
 		i18nextInstance: i18nInstanceType;
 		webkitAudioContext: AudioContext;
