@@ -1,6 +1,6 @@
 import { useSettings } from "@/src/components/Settings/Settings";
 import { type Nullable } from "@/src/types";
-import { MarkerSeverity, type editor } from "@/src/utils/monaco";
+import { type editor, MarkerSeverity } from "@/src/utils/monaco";
 import { cn } from "@/src/utils/utilities";
 import React, { forwardRef } from "react";
 
@@ -16,14 +16,14 @@ const EditorProblems = forwardRef<HTMLDivElement, EditorProblemsProps>(({ classN
 	} = useSettings();
 	const getIcon = (severity: MarkerSeverity) => {
 		switch (severity) {
+			case MarkerSeverity.Error:
+				return "error";
 			case MarkerSeverity.Hint:
 				return "hint";
 			case MarkerSeverity.Info:
 				return "info";
 			case MarkerSeverity.Warning:
 				return "warning";
-			case MarkerSeverity.Error:
-				return "error";
 			default:
 				return "";
 		}
