@@ -3,14 +3,15 @@ import type { configurationId } from "@/src/types";
 import useSectionTitle from "@/src/hooks/useSectionTitle";
 import useSettingsFilter from "@/src/hooks/useSettingsFilter";
 
-import type { CSSEditorProps } from "../../Inputs/CSSEditor/CSSEditor";
 import type { CheckboxProps } from "../../Inputs/CheckBox/CheckBox";
 import type { ColorPickerProps } from "../../Inputs/ColorPicker/ColorPicker";
+import type { CSSEditorProps } from "../../Inputs/CSSEditor/CSSEditor";
 import type { NumberInputProps } from "../../Inputs/Number/Number";
 import type { SelectProps } from "../../Inputs/Select/Select";
 import type { SliderProps } from "../../Inputs/Slider/Slider";
+import type { TextInputProps } from "../../Inputs/TextInput/TextInput";
 
-import { CSSEditor, Checkbox, ColorPicker, NumberInput, Select, Slider } from "../../Inputs";
+import { Checkbox, ColorPicker, CSSEditor, NumberInput, Select, Slider, TextInput } from "../../Inputs";
 
 type SettingInputProps<ID extends configurationId> = {
 	id: ID;
@@ -23,6 +24,7 @@ type SettingInputProps<ID extends configurationId> = {
 	| ({ type: "number" } & NumberInputProps)
 	| ({ type: "select" } & SelectProps<ID>)
 	| ({ type: "slider" } & SliderProps)
+	| ({ type: "text-input" } & TextInputProps)
 );
 function SettingInput<ID extends configurationId>(settingProps: SettingInputProps<ID>) {
 	const { type } = settingProps;
@@ -75,6 +77,10 @@ function SettingInput<ID extends configurationId>(settingProps: SettingInputProp
 			const { className, disabled, id, label, onChange, title, value } = settingProps;
 			return <ColorPicker className={className} disabled={disabled} id={id} label={label} onChange={onChange} title={title} value={value} />;
 		}
+		case "text-input": {
+			const { className, id, input_type, label, onChange, title, value } = settingProps;
+			return <TextInput className={className} id={id} input_type={input_type} label={label} onChange={onChange} title={title} value={value} />;
+		}
 	}
 }
 export default function Setting<ID extends configurationId>(settingProps: SettingInputProps<ID>) {
@@ -86,11 +92,9 @@ export default function Setting<ID extends configurationId>(settingProps: Settin
 			(settingProps.title !== undefined && settingProps.title.toLowerCase().includes(filter.toLowerCase())) ||
 			(settingProps.label !== undefined && settingProps.label.toLowerCase().includes(filter.toLowerCase()))
 		);
-	return (
-		shouldSettingBeVisible && (
+	return shouldSettingBeVisible ?
 			<div className="mx-2 mb-1" title={settingProps.title}>
 				<SettingInput {...settingProps} />
 			</div>
-		)
-	);
+		:	null;
 }
