@@ -4,11 +4,12 @@ import { hideScrollBar } from "./utils";
 
 export async function enableHideScrollBar() {
 	// Wait for the "options" message from the content script
-	const optionsData = await waitForSpecificMessage("options", "request_data", "content");
 	const {
 		data: {
 			options: { enable_hide_scrollbar }
 		}
-	} = optionsData;
-	if (enable_hide_scrollbar) hideScrollBar();
+	} = await waitForSpecificMessage("options", "request_data", "content");
+	// If the hide scroll bar option is disabled, return
+	if (!enable_hide_scrollbar) return;
+	hideScrollBar();
 }
