@@ -1,6 +1,6 @@
 import type { YouTubePlayerDiv, YoutubePlayerQualityLevel } from "@/src/types";
 
-import { browserColorLog, chooseClosestQuality, isShortsPage, isWatchPage, waitForSpecificMessage } from "@/src/utils/utilities";
+import { browserColorLog, chooseClosestQuality, isLivePage, isShortsPage, isWatchPage, waitForSpecificMessage } from "@/src/utils/utilities";
 
 /**
  * Sets the player quality based on the options received from a specific message.
@@ -18,25 +18,19 @@ export default async function setPlayerQuality(): Promise<void> {
 
 	// If automatically set quality option is disabled, return
 	if (!enable_automatically_set_quality) return;
-
 	// If player quality is not specified, return
 	if (!player_quality) return;
-
 	// Get the player element
 	const playerContainer =
-		isWatchPage() ? document.querySelector<YouTubePlayerDiv>("div#movie_player")
+		isWatchPage() || isLivePage() ? document.querySelector<YouTubePlayerDiv>("div#movie_player")
 		: isShortsPage() ? document.querySelector<YouTubePlayerDiv>("div#shorts-player")
 		: null;
-
 	// If player element is not available, return
 	if (!playerContainer) return;
-
 	// If setPlaybackQuality method is not available in the player, return
 	if (!playerContainer.setPlaybackQuality) return;
-
 	// Get the available quality levels
 	const availableQualityLevels = (await playerContainer.getAvailableQualityLevels()) as YoutubePlayerQualityLevel[];
-
 	// Check if the specified player quality is available
 	if (!player_quality || player_quality === "auto") return;
 
