@@ -1,5 +1,6 @@
 import type { YouTubePlayerDiv } from "@/src/types";
 
+import { calculatePlaybackButtonSpeed, updatePlaybackSpeedButtonTooltip } from "@/src/features/playbackSpeedButtons";
 import { setupScrollListeners } from "@/src/features/scrollWheelVolumeControl/utils";
 import OnScreenDisplayManager from "@/src/utils/OnScreenDisplayManager";
 import { isShortsPage, isWatchPage, preventScroll, waitForAllElements, waitForSpecificMessage } from "@/src/utils/utilities";
@@ -31,7 +32,6 @@ export default async function adjustSpeedOnScrollWheel() {
 	// Define the event handler for the scroll wheel events
 	const handleWheel = (event: Event) => {
 		const setOptionsData = async () => (optionsData = await waitForSpecificMessage("options", "request_data", "content"));
-
 		void (async () => {
 			if (!optionsData) return void (await setOptionsData());
 			const {
@@ -44,6 +44,7 @@ export default async function adjustSpeedOnScrollWheel() {
 						osd_display_padding,
 						osd_display_position,
 						osd_display_type,
+						playback_buttons_speed,
 						scroll_wheel_speed_control_modifier_key,
 						speed_adjustment_steps
 					}
@@ -52,14 +53,11 @@ export default async function adjustSpeedOnScrollWheel() {
 			const wheelEvent = event as WheelEvent;
 			// If the modifier key is required and not pressed, return
 			if (enable_scroll_wheel_speed_control && !wheelEvent[scroll_wheel_speed_control_modifier_key]) return void (await setOptionsData());
-
 			// Only prevent default scroll wheel behavior
 			// if we are going to handle the event
 			preventScroll(wheelEvent);
-
 			// Update the options data after preventScroll()
 			await setOptionsData();
-
 			// Get the player element
 			const playerContainer =
 				isWatchPage() ? document.querySelector<YouTubePlayerDiv>("div#movie_player")
@@ -67,11 +65,18 @@ export default async function adjustSpeedOnScrollWheel() {
 				: null;
 			// If player element is not available, return
 			if (!playerContainer) return;
-
 			// Adjust the speed based on the scroll direction
 			const scrollDelta = wheelEvent.deltaY < 0 ? 1 : -1;
 			// Adjust the speed based on the scroll direction and options
-			const { newSpeed } = await adjustSpeed(playerContainer, scrollDelta, speed_adjustment_steps);
+			const { newSpeed } = await adjustSpeed(scrollDelta, speed_adjustment_steps);
+			await updatePlaybackSpeedButtonTooltip(
+				"increasePlaybackSpeedButton",
+				calculatePlaybackButtonSpeed(newSpeed, playback_buttons_speed, "increase")
+			);
+			await updatePlaybackSpeedButtonTooltip(
+				"decreasePlaybackSpeedButton",
+				calculatePlaybackButtonSpeed(newSpeed, playback_buttons_speed, "decrease")
+			);
 			new OnScreenDisplayManager(
 				{
 					displayColor: osd_display_color,
@@ -83,11 +88,14 @@ export default async function adjustSpeedOnScrollWheel() {
 					playerContainer: playerContainer
 				},
 				"yte-osd",
-				{ max: 4, type: "speed", value: newSpeed }
+				{
+					max: 16,
+					type: "speed",
+					value: newSpeed
+				}
 			);
 		})();
 	};
-
 	// Set up the scroll wheel event listeners on the specified container selectors
 	for (const selector of containerSelectors) {
 		setupScrollListeners(selector, handleWheel);
