@@ -10,6 +10,7 @@ import {
 	isWatchPage,
 	round,
 	sendContentMessage,
+	waitForAllElements,
 	waitForSpecificMessage
 } from "@/utils/utilities";
 export async function setupVideoHistory() {
@@ -20,11 +21,9 @@ export async function setupVideoHistory() {
 		}
 	} = await waitForSpecificMessage("options", "request_data", "content");
 	if (!enableVideoHistory) return;
+	if (!isWatchPage()) return;
 	// Get the player container element
-	const playerContainer =
-		isWatchPage() ? document.querySelector<YouTubePlayerDiv>("div#movie_player")
-		: isShortsPage() ? null
-		: null;
+	const playerContainer = document.querySelector<YouTubePlayerDiv>("div#movie_player");
 	// If player container is not available, return
 	if (!playerContainer) return;
 	const playerVideoData = await playerContainer.getVideoData();
@@ -34,7 +33,9 @@ export async function setupVideoHistory() {
 	if (!videoId) return;
 	const videoElement = playerContainer.querySelector<HTMLVideoElement>("video.video-stream.html5-main-video");
 	if (!videoElement) return;
-
+	await waitForAllElements(["#owner #upload-info #channel-name"]);
+	const isOfficialArtistChannel = document.querySelector("#owner #upload-info #channel-name .badge-style-type-verified-artist") !== null;
+	if (isOfficialArtistChannel) return;
 	const videoPlayerTimeUpdateListener = () => {
 		void (async () => {
 			const currentTime = await playerContainer.getCurrentTime();
@@ -58,19 +59,18 @@ export async function promptUserToResumeVideo(cb: () => void) {
 		}
 	} = await waitForSpecificMessage("options", "request_data", "content");
 	if (!enableVideoHistory) return;
-
 	// Get the player container element
 	const playerContainer =
 		isWatchPage() ? document.querySelector<YouTubePlayerDiv>("div#movie_player")
 		: isShortsPage() ? null
 		: null;
-
 	// If player container is not available, return
 	if (!playerContainer) return;
-
+	await waitForAllElements(["#owner #upload-info #channel-name"]);
+	const isOfficialArtistChannel = document.querySelector("#owner #upload-info #channel-name .badge-style-type-verified-artist") !== null;
+	if (isOfficialArtistChannel) return;
 	const { video_id: videoId } = await playerContainer.getVideoData();
 	if (!videoId) return;
-
 	const videoHistoryOneData = await waitForSpecificMessage("videoHistoryOne", "request_data", "content", { id: videoId });
 	if (!videoHistoryOneData) return cb();
 
@@ -93,7 +93,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 	const resumeButtonId = "resume-prompt-button";
 	const promptId = "resume-prompt";
 	const progressBarDuration = 15;
-
 	const prompt = createStyledElement({
 		elementId: promptId,
 		elementType: "div",
@@ -125,7 +124,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 			zIndex: "1000"
 		}
 	});
-
 	const closeButton = createStyledElement({
 		elementId: closeButtonId,
 		elementType: "button",
@@ -160,7 +158,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 		}
 	});
 	resumeButton.textContent = window.i18nextInstance.t("pages.content.features.videoHistory.resumeButton");
-
 	function startCountdown() {
 		if (prompt) prompt.style.display = "block";
 		if (animationFrameId) {
@@ -177,13 +174,11 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 		}
 		animationFrameId = requestAnimationFrame(updateResumeProgress);
 	}
-
 	function hidePrompt() {
 		if (animationFrameId) cancelAnimationFrame(animationFrameId);
 		prompt.style.display = "none";
 		cb();
 	}
-
 	function resumeButtonClickListener() {
 		hidePrompt();
 		browserColorLog(window.i18nextInstance.t("messages.resumingVideo", { VIDEO_TIME: formatTime(videoHistoryEntry.timestamp) }), "FgGreen");
@@ -192,7 +187,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 	}
 
 	if (!elementExists(progressBarId)) prompt.appendChild(progressBar);
-
 	if (!elementExists(closeButtonId)) prompt.appendChild(closeButton);
 
 	const { listener: resumePromptCloseButtonMouseOverListener } = createTooltip({
@@ -203,7 +197,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 	});
 	eventManager.removeEventListener(closeButton, "mouseover", "videoHistory");
 	eventManager.addEventListener(closeButton, "mouseover", resumePromptCloseButtonMouseOverListener, "videoHistory");
-
 	startCountdown();
 
 	const closeListener = () => hidePrompt();
@@ -211,7 +204,6 @@ function createResumePrompt(videoHistoryEntry: VideoHistoryEntry, playerContaine
 	eventManager.addEventListener(resumeButton, "click", resumeButtonClickListener, "videoHistory");
 	eventManager.removeEventListener(closeButton, "click", "videoHistory");
 	eventManager.addEventListener(closeButton, "click", closeListener, "videoHistory");
-
 	// Display the prompt
 	if (elementExists(promptId)) return;
 	prompt.appendChild(resumeButton);
