@@ -6,43 +6,28 @@ import { type RefObject, useEffect } from "react";
 
 const useClickOutside = <ElementType extends HTMLElementTagNameMap[keyof HTMLElementTagNameMap]>(
 	ref: RefObject<ElementType>,
-
 	handler: (event: FocusEvent | MouseEvent | TouchEvent) => void
 ) => {
 	useEffect(() => {
-		let startedInside: Nullable<RefObject<ElementType> | boolean> = false;
-
-		let startedWhenMounted: Nullable<RefObject<ElementType>["current"] | boolean> = false;
-
+		let startedInside: Nullable<boolean | RefObject<ElementType>> = false;
+		let startedWhenMounted: Nullable<boolean | RefObject<ElementType>["current"]> = false;
 		const listener = (event: FocusEvent | MouseEvent | TouchEvent) => {
 			// Do nothing if `mousedown` or `touchstart` started inside ref element
-
 			if (startedInside || !startedWhenMounted) return;
-
 			// Do nothing if clicking ref's element or descendent elements
-
 			if (!ref.current || ref.current.contains(event.target as Node)) return;
-
 			handler(event);
 		};
-
 		const validateEventStart = (event: FocusEvent | MouseEvent | TouchEvent) => {
 			({ current: startedWhenMounted } = ref);
-
 			startedInside = event.target && ref.current && ref.current.contains(event.target as Node);
 		};
-
 		document.addEventListener("mousedown", validateEventStart);
-
 		document.addEventListener("touchstart", validateEventStart);
-
 		document.addEventListener("click", listener);
-
 		return () => {
 			document.removeEventListener("mousedown", validateEventStart);
-
 			document.removeEventListener("touchstart", validateEventStart);
-
 			document.removeEventListener("click", listener);
 		};
 	}, [ref, handler]);
