@@ -1,4 +1,4 @@
-import { type Dispatch, type SetStateAction, createContext } from "react";
+import { createContext, type Dispatch, type SetStateAction } from "react";
 
 export type SettingsFilterContextProps = {
 	filter: string;
