@@ -25,26 +25,26 @@ export const availableLocales = [
 export const localePercentages: Record<AvailableLocales, number> = {
 	"ca-ES": 0,
 	"cs-CZ": 0,
-	"de-DE": 31,
-	"en-GB": 2,
+	"de-DE": 27,
+	"en-GB": 1,
 	"en-US": 100,
-	"es-ES": 54,
+	"es-ES": 48,
 	"fa-IR": 0,
-	"fr-FR": 57,
+	"fr-FR": 50,
 	"he-IL": 0,
 	"hi-IN": 0,
 	"it-IT": 100,
-	"ja-JP": 100,
-	"ko-KR": 94,
+	"ja-JP": 89,
+	"ko-KR": 89,
 	"pl-PL": 0,
-	"pt-BR": 63,
+	"pt-BR": 55,
 	"ru-RU": 100,
-	"sv-SE": 94,
+	"sv-SE": 100,
 	"tr-TR": 69,
-	"uk-UA": 21,
+	"uk-UA": 18,
 	"vi-VN": 0,
 	"zh-CN": 100,
-	"zh-TW": 96
+	"zh-TW": 97
 };
 export const localeDirection: Record<AvailableLocales, "ltr" | "rtl"> = {
 	"ca-ES": "ltr",
