@@ -1,9 +1,13 @@
-/* eslint-disable @typescript-eslint/no-unnecessary-type-assertion */
 import { deepDarkPresets } from "@/src/deepDarkPresets";
-import { type FeatureFuncRecord, featureButtonFunctions } from "@/src/features";
+import { featureButtonFunctions, type FeatureFuncRecord } from "@/src/features";
+import {
+	disableAutomaticallyDisableClosedCaptions,
+	enableAutomaticallyDisableClosedCaptions
+} from "@/src/features/automaticallyDisableClosedCaptions";
 import { enableAutomaticTheaterMode } from "@/src/features/automaticTheaterMode";
 import { featuresInControls } from "@/src/features/buttonPlacement";
 import { getFeatureButton, updateFeatureButtonIcon, updateFeatureButtonTitle } from "@/src/features/buttonPlacement/utils";
+import { addCopyTimestampUrlButton, removeCopyTimestampUrlButton } from "@/src/features/copyTimestampUrlButton";
 import { disableCustomCSS, enableCustomCSS } from "@/src/features/customCSS";
 import { customCSSExists, updateCustomCSS } from "@/src/features/customCSS/utils";
 import { disableDeepDarkCSS, enableDeepDarkCSS } from "@/src/features/deepDarkCSS";
@@ -19,6 +23,10 @@ import {
 	removeHideEndScreenCardsButton
 } from "@/src/features/hideEndScreenCards";
 import { disableHideLiveStreamChat, enableHideLiveStreamChat } from "@/src/features/hideLiveStreamChat";
+import {
+	disableHideOfficialArtistVideosFromHomePage,
+	enableHideOfficialArtistVideosFromHomePage
+} from "@/src/features/hideOfficialArtistVideosFromHomePage";
 import { disableHidePaidPromotionBanner, enableHidePaidPromotionBanner } from "@/src/features/hidePaidPromotionBanner";
 import { enableHideScrollBar } from "@/src/features/hideScrollBar";
 import { hideScrollBar, showScrollBar } from "@/src/features/hideScrollBar/utils";
@@ -34,11 +42,14 @@ import { disablePauseBackgroundPlayers, enablePauseBackgroundPlayers } from "@/s
 import {
 	addDecreasePlaybackSpeedButton,
 	addIncreasePlaybackSpeedButton,
+	calculatePlaybackButtonSpeed,
 	removeDecreasePlaybackSpeedButton,
-	removeIncreasePlaybackSpeedButton
+	removeIncreasePlaybackSpeedButton,
+	updatePlaybackSpeedButtonTooltip
 } from "@/src/features/playbackSpeedButtons";
 import setPlayerQuality from "@/src/features/playerQuality";
 import { restorePlayerSpeed, setPlayerSpeed, setupPlaybackSpeedChangeListener } from "@/src/features/playerSpeed";
+import { disablePlaylistLength, enablePlaylistLength } from "@/src/features/playlistLength";
 import { setupRemainingTime as enableRemainingTime, removeRemainingTimeDisplay } from "@/src/features/remainingTime";
 import enableRememberVolume from "@/src/features/rememberVolume";
 import enableRemoveRedirect from "@/src/features/removeRedirect";
@@ -57,20 +68,20 @@ import volumeBoost, {
 	removeVolumeBoostButton
 } from "@/src/features/volumeBoost";
 import { i18nService } from "@/src/i18n";
-import { type ToggleFeatures, type ToggleIcon, getFeatureIcon, toggleFeatures } from "@/src/icons";
+import { getFeatureIcon, type ToggleFeatures, toggleFeatures, type ToggleIcon } from "@/src/icons";
 import {
 	type AllButtonNames,
 	type ButtonPlacement,
 	type ExtensionSendOnlyMessageMappings,
 	type FeatureToMultiButtonMap,
+	featureToMultiButtonsMap,
 	type KeysOfUnion,
 	type Messages,
 	type MultiButtonFeatureNames,
 	type MultiButtonNames,
 	type SingleButtonFeatureNames,
 	type SingleButtonNames,
-	type YouTubePlayerDiv,
-	featureToMultiButtonsMap
+	type YouTubePlayerDiv
 } from "@/src/types";
 import eventManager from "@/utils/EventManager";
 import {
@@ -78,7 +89,9 @@ import {
 	findKeyByValue,
 	formatError,
 	groupButtonChanges,
+	isLivePage,
 	isNewYouTubeVideoLayout,
+	isPlaylistPage,
 	isShortsPage,
 	isWatchPage,
 	sendContentOnlyMessage,
@@ -134,7 +147,9 @@ const alwaysShowProgressBar = async function () {
 	progressPlay += progressWidth;
 	progressLoad += progressWidth;
 };
-
+function shouldEnableFeaturesFuncReturn() {
+	return !(isWatchPage() || isShortsPage() || isPlaylistPage() || isLivePage());
+}
 const enableFeatures = () => {
 	browserColorLog(`Enabling features...`, "FgMagenta");
 	void (async () => {
@@ -153,15 +168,13 @@ const enableFeatures = () => {
 			enableShareShortener(),
 			enableSkipContinueWatching(),
 			enablePauseBackgroundPlayers(),
-			enableRememberVolume(),
 			enableHideScrollBar(),
 			enableCustomCSS(),
-			enableDeepDarkCSS()
+			enableDeepDarkCSS(),
+			enableHideOfficialArtistVideosFromHomePage()
 		]);
-
 		// Use a guard clause to reduce amount of times nesting code happens
-		if (!(isWatchPage() || isShortsPage())) return;
-
+		if (shouldEnableFeaturesFuncReturn()) return;
 		void Promise.all([
 			promptUserToResumeVideo(() => void setupVideoHistory()),
 			setupPlaybackSpeedChangeListener(),
@@ -171,13 +184,14 @@ const enableFeatures = () => {
 			enableRememberVolume(),
 			enableAutomaticTheaterMode(),
 			enableRemainingTime(),
-			volumeBoost(),
 			setPlayerQuality(),
 			setPlayerSpeed(),
 			adjustVolumeOnScrollWheel(),
 			adjustSpeedOnScrollWheel(),
 			enableHideTranslateComment(),
-			enableHideEndScreenCards()
+			enableHideEndScreenCards(),
+			enablePlaylistLength(),
+			enableAutomaticallyDisableClosedCaptions()
 		]);
 		// Enable feature menu before calling button functions
 		await enableFeatureMenu();
@@ -224,6 +238,8 @@ const enableFeatures = () => {
 		await openTranscriptButton();
 		await addMaximizePlayerButton();
 		await addLoopButton();
+		await addCopyTimestampUrlButton();
+		await volumeBoost();
 	})();
 };
 const getFeatureFunctions = (featureName: AllButtonNames, oldPlacement: ButtonPlacement) => {
@@ -232,17 +248,18 @@ const getFeatureFunctions = (featureName: AllButtonNames, oldPlacement: ButtonPl
 	if (!featureFunctions) {
 		throw new Error(`Feature '${featureName}' not found in featureButtonFunctions`);
 	}
-
 	// Cast featureFunctions to FeatureFuncRecord
 	const castFeatureFunctions = featureFunctions as unknown as FeatureFuncRecord;
-
 	return {
 		add: () => castFeatureFunctions.add(),
 		remove: () => castFeatureFunctions.remove(oldPlacement)
 	};
 };
 function handleSoftNavigate() {
-	// Listen to YouTube's soft navigate event
+	// Remove existing listeners
+	document.removeEventListener("yt-navigate-finish", enableFeatures);
+	document.removeEventListener("yt-player-updated", enableFeatures);
+	// Add listeners
 	document.addEventListener("yt-navigate-finish", enableFeatures);
 	document.addEventListener("yt-player-updated", enableFeatures);
 }
@@ -258,7 +275,7 @@ window.addEventListener("DOMContentLoaded", function () {
 			enableFeatures();
 			handleSoftNavigate();
 		} else if (!isFirstLoad) {
-			handleSoftNavigate;
+			handleSoftNavigate();
 		}
 		isFirstLoad = false;
 		/**
@@ -281,6 +298,34 @@ window.addEventListener("DOMContentLoaded", function () {
 				}
 				if (!message) return;
 				switch (message.type) {
+					case "hideOfficialArtistVideosFromHomePageChange": {
+						const {
+							data: { hideOfficialArtistVideosFromHomePageEnabled }
+						} = message;
+						if (hideOfficialArtistVideosFromHomePageEnabled) {
+							await enableHideOfficialArtistVideosFromHomePage();
+						} else {
+							disableHideOfficialArtistVideosFromHomePage();
+						}
+						break;
+					}
+					case "playlistLengthChange": {
+						const {
+							data: { playlistLengthEnabled }
+						} = message;
+						if (playlistLengthEnabled) {
+							await enablePlaylistLength();
+						} else {
+							disablePlaylistLength();
+						}
+						break;
+					}
+					case "playlistLengthGetMethodChange":
+					case "playlistWatchTimeGetMethodChange": {
+						disablePlaylistLength();
+						await enablePlaylistLength();
+						break;
+					}
 					case "volumeBoostChange": {
 						const {
 							data: { volumeBoostEnabled, volumeBoostMode }
@@ -303,7 +348,6 @@ window.addEventListener("DOMContentLoaded", function () {
 						const {
 							data: { volumeBoostAmount, volumeBoostEnabled, volumeBoostMode }
 						} = message;
-
 						switch (volumeBoostMode) {
 							case "global": {
 								if (volumeBoostEnabled) applyVolumeBoost(volumeBoostAmount);
@@ -320,10 +364,34 @@ window.addEventListener("DOMContentLoaded", function () {
 					}
 					case "playerSpeedChange": {
 						const {
-							data: { enableForcedPlaybackSpeed, playerSpeed }
-						} = message;
-						if (enableForcedPlaybackSpeed && playerSpeed) await setPlayerSpeed(Number(playerSpeed));
-						else if (!enableForcedPlaybackSpeed) restorePlayerSpeed();
+							data: {
+								options: { playback_buttons_speed: playbackSpeedPerClick }
+							}
+						} = await waitForSpecificMessage("options", "request_data", "content");
+						if (enableForcedPlaybackSpeed && playerSpeed) {
+							await updatePlaybackSpeedButtonTooltip(
+								"increasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "increase")
+							);
+							await updatePlaybackSpeedButtonTooltip(
+								"decreasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(playerSpeed, playbackSpeedPerClick, "decrease")
+							);
+							await setPlayerSpeed(Number(playerSpeed));
+						} else if (!enableForcedPlaybackSpeed) {
+							restorePlayerSpeed();
+							const videoElement = document.querySelector<HTMLVideoElement>("video");
+							if (!videoElement) return;
+							const { playbackRate: currentSpeed } = videoElement;
+							await updatePlaybackSpeedButtonTooltip(
+								"increasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "increase")
+							);
+							await updatePlaybackSpeedButtonTooltip(
+								"decreasePlaybackSpeedButton",
+								calculatePlaybackButtonSpeed(currentSpeed, playbackSpeedPerClick, "decrease")
+							);
+						}
 						break;
 					}
 					case "screenshotButtonChange": {
@@ -408,6 +476,17 @@ window.addEventListener("DOMContentLoaded", function () {
 						} = message;
 						if (loopButtonEnabled) await addLoopButton();
 						else await removeLoopButton();
+						break;
+					}
+					case "copyTimestampUrlButtonChange": {
+						const {
+							data: { copyTimestampUrlButtonEnabled }
+						} = message;
+						if (copyTimestampUrlButtonEnabled) {
+							await addCopyTimestampUrlButton();
+						} else {
+							await removeCopyTimestampUrlButton();
+						}
 						break;
 					}
 					case "forwardRewindButtonsChange": {
@@ -533,6 +612,17 @@ window.addEventListener("DOMContentLoaded", function () {
 						else await removeHideEndScreenCardsButton();
 						break;
 					}
+					case "automaticallyDisableClosedCaptionsChange": {
+						const {
+							data: { automaticallyDisableClosedCaptionsEnabled }
+						} = message;
+						if (automaticallyDisableClosedCaptionsEnabled) {
+							await enableAutomaticallyDisableClosedCaptions();
+						} else {
+							await disableAutomaticallyDisableClosedCaptions();
+						}
+						break;
+					}
 					case "hideLiveStreamChatChange": {
 						const {
 							data: { hideLiveStreamChatEnabled }
@@ -644,7 +734,6 @@ window.addEventListener("DOMContentLoaded", function () {
 						// If the size button is not available return
 						if (!sizeButton) return;
 						sizeButton.click();
-
 						break;
 					}
 					case "featureMenuOpenTypeChange": {
@@ -825,7 +914,6 @@ window.onbeforeunload = function () {
 window.addEventListener("error", (event) => {
 	event.preventDefault();
 	const {
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 		error: { stack: errorLine }
 	} = event;
 	browserColorLog(formatError(event.error) + "\nAt: " + errorLine, "FgRed");
