@@ -1,5 +1,5 @@
-import Options from "@/pages/options/Options";
 import "@/pages/options/index.css";
+import Options from "@/pages/options/Options";
 import React from "react";
 import { createRoot } from "react-dom/client";
 
