@@ -1,6 +1,6 @@
 import "@/assets/styles/tailwind.css";
-import Popup from "@/pages/popup/Popup";
 import "@/pages/popup/index.css";
+import Popup from "@/pages/popup/Popup";
 import React from "react";
 import { createRoot } from "react-dom/client";
 
