@@ -7,7 +7,7 @@ import z, { ZodType } from "zod";
 import type { DeepDarkPreset } from "../deepDarkPresets";
 import type { AvailableLocales } from "../i18n/constants";
 // #region Utility types
-export type Nullable<T> = T | null;
+export type Nullable<T> = null | T;
 export type NonNullable<T> = T extends Nullable<T> ? Exclude<T, null> : T;
 export type NonNullableObject<T> = { [K in keyof T]: NonNullable<T[K]> };
 export type AnyFunction = (...args: any[]) => void;
@@ -61,7 +61,7 @@ type PathImpl<T, Key extends keyof T> =
 			:	`${Key}.${PathImpl<T[Key], keyof T[Key]>}` | Key
 		:	Key
 	:	never;
-export type Path<T> = PathImpl<T, keyof T> | keyof T;
+export type Path<T> = keyof T | PathImpl<T, keyof T>;
 export type PathValue<T, P extends Path<T>> =
 	P extends `${infer Key}.${infer Rest}` ?
 		Key extends keyof T ?
@@ -107,9 +107,9 @@ export const youtubePlayerQualityLevels = [
 export type YoutubePlayerQualityLevel = (typeof youtubePlayerQualityLevels)[number];
 export const PlayerQualityFallbackStrategy = ["higher", "lower"] as const;
 export type PlayerQualityFallbackStrategy = (typeof PlayerQualityFallbackStrategy)[number];
-export const youtubePlayerSpeedRatesExtended = [2.25, 2.5, 2.75, 3, 3.25, 3.75, 4] as const;
-export const youtubePlayerSpeedRates = [0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2, ...youtubePlayerSpeedRatesExtended] as const;
-export const youtubePlaybackSpeedButtonsRates = [0.25, 0.5, 0.75, 1] as const;
+export const youtubePlayerMaxSpeed = 16;
+export const youtubePlayerMinSpeed = 0.07;
+export const youtubePlayerSpeedStep = 0.01;
 export const screenshotTypes = ["file", "clipboard"] as const;
 export type ScreenshotType = (typeof screenshotTypes)[number];
 export const screenshotFormats = ["png", "jpeg", "webp"] as const;
@@ -122,6 +122,10 @@ export const videoHistoryResumeTypes = ["automatic", "prompt"] as const;
 export type VideoHistoryResumeType = (typeof videoHistoryResumeTypes)[number];
 export const buttonPlacements = ["below_player", "feature_menu", "player_controls_left", "player_controls_right"] as const;
 export type ButtonPlacement = (typeof buttonPlacements)[number];
+export const playlistWatchTimeGetMethod = ["duration", "youtube"] as const;
+export type PlaylistWatchTimeGetMethod = (typeof playlistWatchTimeGetMethod)[number];
+export const playlistLengthGetMethod = ["api", "html"] as const;
+export type PlaylistLengthGetMethod = (typeof playlistLengthGetMethod)[number];
 export const featureMenuOpenTypes = ["click", "hover"] as const;
 export type MultiButtonChange = {
 	[K in MultiButtonFeatureNames]: {
@@ -177,7 +181,7 @@ const featureToMultiButtonMapEntries: FeatureToMultiButtonMap = {
 export const featureToMultiButtonsMap = new Map(
 	Object.keys(featureToMultiButtonMapEntries).map((key) => [
 		key,
-		// eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+
 		Object.keys(featureToMultiButtonMapEntries[key]) as KeysOfUnion<FeatureToMultiButtonMap[typeof key]>[]
 	])
 );
@@ -185,6 +189,7 @@ export type FeatureMenuItemIconId = `yte-${AllButtonNames}-icon`;
 export type FeatureMenuItemId = `yte-feature-${AllButtonNames}-menuitem`;
 export type FeatureMenuItemLabelId = `yte-${AllButtonNames}-label`;
 export const buttonNames = Object.keys({
+	copyTimestampUrlButton: "",
 	decreasePlaybackSpeedButton: "",
 	forwardButton: "",
 	hideEndScreenCardsButton: "",
@@ -197,6 +202,7 @@ export const buttonNames = Object.keys({
 	volumeBoostButton: ""
 } satisfies Record<AllButtonNames, "">);
 export const buttonNameToSettingName = {
+	copyTimestampUrlButton: "enable_copy_timestamp_url_button",
 	decreasePlaybackSpeedButton: "enable_playback_speed_buttons",
 	forwardButton: "enable_forward_rewind_buttons",
 	hideEndScreenCardsButton: "enable_hide_end_screen_cards_button",
@@ -231,6 +237,37 @@ export type Notification = {
 	removeAfterMs?: number;
 	timestamp?: number;
 	type: NotificationType;
+};
+export type YouTubePlaylistResponse = {
+	items: YouTubePlaylistItem[];
+	nextPageToken?: string;
+};
+export type YouTubePlaylistItem = {
+	contentDetails: {
+		videoId: string;
+	};
+};
+
+export type YouTubeVideoResponse = {
+	items: YouTubeVideoItem[];
+};
+
+export type YouTubeVideoItem = {
+	contentDetails: {
+		duration: string; // ISO 8601 duration format
+	};
+};
+export type YouTubeAPIQuotaError = {
+	error: {
+		code: number;
+		errors: { domain: string; message: string; reason: string }[];
+		message: string;
+	};
+};
+export type VideoDetails = {
+	duration: number;
+	progress: number;
+	videoId: Nullable<string>;
 };
 export type CrowdinLanguageProgressResponse = {
 	data: {
@@ -316,8 +353,13 @@ export type ContentToBackgroundSendOnlyMessageMappings = {
 	pauseBackgroundPlayers: ActionMessage<"pauseBackgroundPlayers">;
 };
 export type ExtensionSendOnlyMessageMappings = {
+	automaticallyDisableClosedCaptionsChange: DataResponseMessage<
+		"automaticallyDisableClosedCaptionsChange",
+		{ automaticallyDisableClosedCaptionsEnabled: boolean }
+	>;
 	automaticTheaterModeChange: DataResponseMessage<"automaticTheaterModeChange", { automaticTheaterModeEnabled: boolean }>;
 	buttonPlacementChange: DataResponseMessage<"buttonPlacementChange", ButtonPlacementChange>;
+	copyTimestampUrlButtonChange: DataResponseMessage<"copyTimestampUrlButtonChange", { copyTimestampUrlButtonEnabled: boolean }>;
 	customCSSChange: DataResponseMessage<"customCSSChange", { customCSSCode: string; customCSSEnabled: boolean }>;
 	deepDarkThemeChange: DataResponseMessage<
 		"deepDarkThemeChange",
@@ -331,6 +373,10 @@ export type ExtensionSendOnlyMessageMappings = {
 		{ hideEndScreenCardsButtonPlacement: ButtonPlacement; hideEndScreenCardsEnabled: boolean }
 	>;
 	hideLiveStreamChatChange: DataResponseMessage<"hideLiveStreamChatChange", { hideLiveStreamChatEnabled: boolean }>;
+	hideOfficialArtistVideosFromHomePageChange: DataResponseMessage<
+		"hideOfficialArtistVideosFromHomePageChange",
+		{ hideOfficialArtistVideosFromHomePageEnabled: boolean }
+	>;
 	hidePaidPromotionBannerChange: DataResponseMessage<"hidePaidPromotionBannerChange", { hidePaidPromotionBannerEnabled: boolean }>;
 	hideScrollBarChange: DataResponseMessage<"hideScrollBarChange", { hideScrollBarEnabled: boolean }>;
 	hideShortsChange: DataResponseMessage<"hideShortsChange", { hideShortsEnabled: boolean }>;
@@ -351,6 +397,9 @@ export type ExtensionSendOnlyMessageMappings = {
 		{ playbackButtonsSpeed: number; playbackSpeedButtonsEnabled: boolean }
 	>;
 	playerSpeedChange: DataResponseMessage<"playerSpeedChange", { enableForcedPlaybackSpeed: boolean; playerSpeed?: number }>;
+	playlistLengthChange: DataResponseMessage<"playlistLengthChange", { playlistLengthEnabled: boolean }>;
+	playlistLengthGetMethodChange: DataResponseMessage<"playlistLengthGetMethodChange", undefined>;
+	playlistWatchTimeGetMethodChange: DataResponseMessage<"playlistWatchTimeGetMethodChange", undefined>;
 	remainingTimeChange: DataResponseMessage<"remainingTimeChange", { remainingTimeEnabled: boolean }>;
 	rememberVolumeChange: DataResponseMessage<"rememberVolumeChange", { rememberVolumeEnabled: boolean }>;
 	removeRedirectChange: DataResponseMessage<"removeRedirectChange", { removeRedirectEnabled: boolean }>;
@@ -365,6 +414,7 @@ export type ExtensionSendOnlyMessageMappings = {
 		}
 	>;
 	skipContinueWatchingChange: DataResponseMessage<"skipContinueWatchingChange", { skipContinueWatchingEnabled: boolean }>;
+	timestampPeekChange: DataResponseMessage<"timestampPeekChange", { timestampPeekEnabled: boolean }>;
 	videoHistoryChange: DataResponseMessage<"videoHistoryChange", { videoHistoryEnabled: boolean }>;
 	volumeBoostAmountChange: DataResponseMessage<
 		"volumeBoostAmountChange",
@@ -408,7 +458,9 @@ export type configuration = {
 	deep_dark_custom_theme_colors: DeepDarkCustomThemeColors;
 	deep_dark_preset: DeepDarkPreset;
 	enable_automatic_theater_mode: boolean;
+	enable_automatically_disable_closed_captions: boolean;
 	enable_automatically_set_quality: boolean;
+	enable_copy_timestamp_url_button: boolean;
 	enable_custom_css: boolean;
 	enable_deep_dark_theme: boolean;
 	enable_forced_playback_speed: boolean;
@@ -416,6 +468,7 @@ export type configuration = {
 	enable_hide_end_screen_cards: boolean;
 	enable_hide_end_screen_cards_button: boolean;
 	enable_hide_live_stream_chat: boolean;
+	enable_hide_official_artist_videos_from_home_page: boolean;
 	enable_hide_paid_promotion_banner: boolean;
 	enable_hide_scrollbar: boolean;
 	enable_hide_shorts: boolean;
@@ -426,6 +479,7 @@ export type configuration = {
 	enable_open_youtube_settings_on_hover: boolean;
 	enable_pausing_background_players: boolean;
 	enable_playback_speed_buttons: boolean;
+	enable_playlist_length: boolean;
 	enable_redirect_remover: boolean;
 	enable_remaining_time: boolean;
 	enable_remember_last_volume: boolean;
@@ -437,6 +491,7 @@ export type configuration = {
 	enable_share_shortener: boolean;
 	enable_shorts_auto_scroll: boolean;
 	enable_skip_continue_watching: boolean;
+	enable_timestamp_peek: boolean;
 	enable_video_history: boolean;
 	enable_volume_boost: boolean;
 	feature_menu_open_type: FeatureMenuOpenType;
@@ -452,6 +507,8 @@ export type configuration = {
 	player_quality: YoutubePlayerQualityLevel;
 	player_quality_fallback_strategy: PlayerQualityFallbackStrategy;
 	player_speed: number;
+	playlist_length_get_method: PlaylistLengthGetMethod;
+	playlist_watch_time_get_method: PlaylistWatchTimeGetMethod;
 	remembered_volumes: RememberedVolumes;
 	screenshot_format: ScreenshotFormat;
 	screenshot_save_as: ScreenshotType;
@@ -462,6 +519,7 @@ export type configuration = {
 	volume_adjustment_steps: number;
 	volume_boost_amount: number;
 	volume_boost_mode: VolumeBoostMode;
+	youtube_data_api_v3_key: string;
 };
 export type configurationKeys = keyof configuration;
 export type configurationId = Path<configuration>;
