@@ -8,4 +8,4 @@ import { MarkerSeverity } from "monaco-editor/esm/vs/editor/editor.api";
 import * as monaco from "monaco-editor/esm/vs/editor/editor.api";
 import "monaco-editor/esm/vs/language/css/cssMode";
 import "monaco-editor/esm/vs/language/css/monaco.contribution";
-export { MarkerSeverity, type editor, monaco };
+export { type editor, MarkerSeverity, monaco };
