@@ -6,14 +6,15 @@ import type {
 	AllButtonNames,
 	AnyFunction,
 	ButtonPlacementChange,
+	configuration,
 	ContentSendOnlyMessageMappings,
 	ContentToBackgroundSendOnlyMessageMappings,
 	DeepPartial,
 	ExtensionSendOnlyMessageMappings,
 	FeatureToMultiButtonMap,
 	MessageMappings,
-	MessageSource,
 	Messages,
+	MessageSource,
 	MultiButtonChange,
 	Nullable,
 	OnScreenDisplayPosition,
@@ -25,13 +26,12 @@ import type {
 	SingleButtonChange,
 	SingleButtonFeatureNames,
 	SingleButtonNames,
-	YoutubePlayerQualityLevel,
-	configuration
+	YoutubePlayerQualityLevel
 } from "../types";
 import type { SVGElementAttributes } from "./SVGElementAttributes";
 
 import { buttonNameToSettingName, featureToMultiButtonsMap, youtubePlayerQualityLevels } from "../types";
-import { type FeatureName, eventManager } from "./EventManager";
+import { eventManager, type FeatureName } from "./EventManager";
 
 export const isStrictEqual = (value1: unknown) => (value2: unknown) => value1 === value2;
 export const isNotStrictEqual = (value1: unknown) => (value2: unknown) => value1 !== value2;
@@ -47,8 +47,7 @@ export const removeSpecialCharacters = (value: string) => {
 export const unique = (values: string[]) => [...new Set(values)];
 
 export const clamp = (value: number, min: number, max: number) => Math.min(Math.max(value, min), max);
-
-export const round = (value: number, decimals = 0) => Number(`${Math.round(Number(`${value}e${decimals}`))}e-${decimals}`);
+export const round = (value: number, decimals = 0) => Number(`${Math.round(Number(`${value + Number.EPSILON}e${decimals}`))}e-${decimals}`);
 
 export const toDivisible = (value: number, divider: number): number => Math.ceil(value / divider) * divider;
 
@@ -178,7 +177,7 @@ function groupMessages(messages: { message: string; styling: string[] }[]): Arra
  * @returns The colorized log message.
  */
 export function browserColorLog(message: string, type?: ColorType) {
-	const prependLog = colorizeLog(`[YouTube Enhancer]`, "FgCyan");
+	const prependLog = colorizeLog(`[${getFormattedTimestamp()}] [YouTube Enhancer]`, "FgCyan");
 	const colorizedMessage = colorizeLog(message, type);
 	console.log(...groupMessages([prependLog, colorizedMessage]));
 }
@@ -365,6 +364,14 @@ export function isShortsPage() {
 	const firstSection = extractFirstSectionFromYouTubeURL(window.location.href);
 	return firstSection === "shorts";
 }
+export function isPlaylistPage() {
+	const firstSection = extractFirstSectionFromYouTubeURL(window.location.href);
+	return firstSection === "playlist";
+}
+export function isLivePage() {
+	const firstSection = extractFirstSectionFromYouTubeURL(window.location.href);
+	return firstSection === "live";
+}
 export function formatError(error: unknown) {
 	if (error instanceof Error) {
 		return `${error.message}\n${error?.stack}`;
@@ -476,7 +483,7 @@ export function parseStoredValue(value: string) {
 		if (typeof parsedValue === "boolean" || typeof parsedValue === "number" || typeof parsedValue === "object") {
 			return parsedValue; // Return the parsed value
 		}
-	} catch (error) {
+	} catch (_) {
 		// If parsing or type checking fails, return the original value as a string
 	}
 	// If parsing or type checking fails, return the original value as a string
@@ -603,7 +610,7 @@ export function createStyledElement<ID extends string, K extends keyof HTMLEleme
 	// Return the element
 	return element;
 }
-type SVGChildElement = SVGElement | SVGPathElement | SVGTSpanElement | SVGTextElement;
+type SVGChildElement = SVGElement | SVGPathElement | SVGTextElement | SVGTSpanElement;
 
 export function createSVGElement<K extends keyof SVGElementTagNameMap>(
 	tagName: K,
@@ -633,12 +640,6 @@ export function calculateCanvasPosition(displayPosition: OnScreenDisplayPosition
 	let styles: Partial<CSSStyleDeclaration> = {};
 
 	switch (displayPosition) {
-		case "top_left":
-			styles = { left: `${displayPadding}px`, top: `${displayPadding + paddingTop}px` };
-			break;
-		case "top_right":
-			styles = { right: `${displayPadding}px`, top: `${displayPadding + paddingTop}px` };
-			break;
 		case "bottom_left":
 			styles = { bottom: `${displayPadding + paddingBottom}px`, left: `${displayPadding}px` };
 			break;
@@ -647,6 +648,12 @@ export function calculateCanvasPosition(displayPosition: OnScreenDisplayPosition
 			break;
 		case "center":
 			styles = { left: "50%", top: "50%", transform: "translate(-50%, -50%)" };
+			break;
+		case "top_left":
+			styles = { left: `${displayPadding}px`, top: `${displayPadding + paddingTop}px` };
+			break;
+		case "top_right":
+			styles = { right: `${displayPadding}px`, top: `${displayPadding + paddingTop}px` };
 			break;
 		default:
 			console.error("Invalid display position");
@@ -670,7 +677,6 @@ export function getPathValue<T, P extends Path<T>>(obj: T, path: P): PathValue<T
 
 	for (const key of keys) {
 		if (value && typeof value === "object" && key in value) {
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
 			({ [key]: value } = value);
 		} else {
 			console.error(`Invalid path: ${String(path)}`);
@@ -765,4 +771,71 @@ export function isNewYouTubeVideoLayout(): boolean {
 	} else {
 		return false; // It's the old layout
 	}
+}
+export function getFormattedTimestamp() {
+	const now = new Date();
+
+	const month = (now.getMonth() + 1).toString().padStart(2, "0");
+	const day = now.getDate().toString().padStart(2, "0");
+	const year = now.getFullYear().toString().substr(-2);
+	const hours = now.getHours();
+	const minutes = now.getMinutes().toString().padStart(2, "0");
+	const seconds = now.getSeconds().toString().padStart(2, "0");
+	const milliseconds = now.getMilliseconds().toString().padStart(3, "0");
+
+	const period = hours >= 12 ? "PM" : "AM";
+	const paddedHours = (hours % 12 || 12).toString().padStart(2, "0"); // Convert to 12-hour format and handle midnight (0 hours)
+
+	return `${month}/${day}/${year} ${paddedHours}:${minutes}:${seconds}:${milliseconds} ${period}`;
+}
+/**
+ * Parses an ISO 8601 duration string and returns the total number of seconds.
+ *
+ * @param {string} duration - The ISO 8601 duration string to parse.
+ * @return {number} The total number of seconds represented by the duration string.
+ */
+export function parseISO8601Duration(duration: string): number {
+	// Regular expression to match ISO 8601 duration format
+	const regex = /^PT(?:(\d+)H)?(?:(\d+)M)?(?:(\d+)S)?$/;
+	// Extract hours, minutes, and seconds from the duration string
+	const matches = regex.exec(duration);
+	// If the duration string does not match the expected format, return 0
+	if (!matches) return 0;
+
+	// Parse the hours, minutes, and seconds from the matches array
+	const hours = parseInt(matches[1] || "0", 10);
+	const minutes = parseInt(matches[2] || "0", 10);
+	const seconds = parseInt(matches[3] || "0", 10);
+
+	// Calculate the total number of seconds by multiplying hours, minutes, and seconds
+	return hours * 3600 + minutes * 60 + seconds;
+}
+
+/**
+ * Formats a duration in seconds into a string representation.
+ *
+ * @param {number} seconds - The duration in seconds.
+ * @return {string} The formatted duration string in the format "HHhMMmSSs".
+ */
+export function formatDuration(seconds: number): string {
+	// Calculate the hours, minutes, and seconds
+	const hours = Math.floor(seconds / 3600);
+	const minutes = Math.floor((seconds % 3600) / 60);
+	const secs = seconds % 60;
+
+	// Format the hours, minutes, and seconds with leading zeros
+	const formattedHours = hours.toString();
+	const formattedMinutes = minutes.toString().padStart(2, "0");
+	const formattedSeconds = secs.toString().padStart(2, "0");
+
+	// Combine the formatted values into a single string
+	return `${formattedHours}:${formattedMinutes}:${formattedSeconds}`;
+}
+export function timeStringToSeconds(timeString: string): number {
+	const parts = timeString.split(":").reverse();
+	let seconds = 0;
+	for (let i = 0; i < parts.length; i++) {
+		seconds += parseInt(parts[i], 10) * Math.pow(60, i);
+	}
+	return seconds;
 }
